+++ test/Veggerby.Boards.Tests/Core/GameEngineTests.cs
@@ -21,7 +21,7 @@ public class GameEngineTests
             var gamePhaseRoot = GamePhase.New(1, "test", new NullGameStateCondition(), GameEventRule<IGameEvent>.Null);
 
             // act
-            var actual = new GameEngine(game, gamePhaseRoot);
+            var actual = new GameEngine(game, gamePhaseRoot, builder.Engine.DecisionPlan);
 
             // assert
             actual.Should().NotBeNull();
@@ -36,7 +36,8 @@ public class GameEngineTests
             var gamePhaseRoot = GamePhase.New(1, "test", new NullGameStateCondition(), GameEventRule<IGameEvent>.Null);
 
             // act
-            var actual = () => new GameEngine(null, gamePhaseRoot);
+            var dummyPlan = Veggerby.Boards.Flows.DecisionPlan.DecisionPlan.Compile(gamePhaseRoot);
+            var actual = () => new GameEngine(null, gamePhaseRoot, dummyPlan);
 
             // assert
             actual.Should().Throw<ArgumentNullException>().WithParameterName("game");
@@ -50,7 +51,7 @@ public class GameEngineTests
             var game = builder.Game;
 
             // act
-            var actual = () => new GameEngine(game, null);
+            var actual = () => new GameEngine(game, null, builder.Engine.DecisionPlan);
 
             // assert
             actual.Should().Throw<ArgumentNullException>().WithParameterName("gamePhaseRoot");